% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_FINESS.R
\name{clean_FINESS}
\alias{clean_FINESS}
\title{FINESS code cleaning}
\usage{
clean_FINESS(char)
}
\arguments{
\item{char}{A character vector containing FINESS codes with missing leading "0". Each code should be 8 or 9 characters.}
}
\value{
A character vector containing the FINESS code with a leading "0" when needed
}
\description{
Adds the leading "0" to FINESS codes if it contains only 8 caracters instead of 9.
}
\examples{
library(rUrgAra)
clean_FINESS(c("10784320", "380000174"))
}
