% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exemple-dataset.R
\docType{data}
\name{df_ex_charge}
\alias{df_ex_charge}
\title{df_ex_charge}
\format{
A tibble frame with 500 rows and 2 variables:
\describe{
\item{ ENTREE }{  dttm }
\item{ SORTIE }{  dttm }
\item{ Etablissement }{ character}
}
}
\usage{
df_ex_charge
}
\description{
exemple dataset containing a set of 500 entry and exit date in january 2022
}
\keyword{datasets}
