% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_delais.R
\name{fct_calc_delay}
\alias{fct_calc_delay}
\title{fct_calc_delay}
\usage{
fct_calc_delay(base, d1, d2, lower_bound, upper_bound, min_display)
}
\arguments{
\item{base}{Table contenant les date/heure}

\item{d1}{Nom de premier date/heure}

\item{d2}{Nom de deuxieme date/heure}

\item{lower_bound}{Borne minimale du délais. Les valeurs inférieurs sont supprimées}

\item{upper_bound}{Borne maximale du délais. Les valeurs supérieurs sont supprimées}

\item{min_display}{Le nombre minimal de patient de délai valide pour afficher
la médiane et l'interquartile. Remplacé par "X" sinon.}
}
\value{
Une liste de 4 valeurs (lab_del/Q1_del/med_del/Q3_del)
}
\description{
Calcul les délais entre deux dates/heures dans une base et supprime les valeurs hors délais
}
