% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fct_f_big}
\alias{fct_f_big}
\title{Format grand nombres}
\usage{
fct_f_big(val)
}
\arguments{
\item{val}{Numeric}
}
\value{
Une chaine de caractère contenant le nombre donné en entrée avec un espace en séparateur des milliers
}
\description{
Format grand nombres
}
