% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{fct_format_control}
\alias{fct_format_control}
\title{Controle values in a variable}
\usage{
fct_format_control(var, val_autor, lab, replacement = "NC")
}
\arguments{
\item{var}{Variable to check}

\item{val_autor}{Vector of authorized values}

\item{lab}{The name of the variable (only used to print the number of removed values)}

\item{replacement}{Value to replace "Non conform" modalities with. "NC" is used by default.}
}
\value{
A vector with NA left as NA and values outside of "val_autor" transformed into "NC" (Not conform)
}
\description{
Controle values in a variable
}
\examples{
var = c(NA, LETTERS[1:5])
fct_format_control(var, val_autor = c("A", "C", "F"), lab = "lettres")

}
