% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_add_logo.R
\name{fct_make_icon_qual}
\alias{fct_make_icon_qual}
\title{Icon qualité des données}
\usage{
fct_make_icon_qual(p)
}
\arguments{
\item{p}{Numeric entre 0 et 100 => Nombre à insérer au centre du cercle}
}
\value{
un ggplot
}
\description{
Icon qualité des données
}
