% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_p_value}
\alias{format_p_value}
\title{format_p_value}
\usage{
format_p_value(p, digits = 3, decimal.mark = ".")
}
\arguments{
\item{p}{numeric ou vecteur de numeric contenant des p-valeur (0 à 1)}

\item{digits}{numeric : Nombre de chiffres après la vircule}

\item{decimal.mark}{character : Séparateur des décimales}
}
\value{
character ou vecteur de characters
}
\description{
Mets en forme une p-value en format texte.
}
\examples{
library(rUrgAra)
format_p_value(NA_real_)
format_p_value(1)
format_p_value(0.52, decimal.mark = ",")
format_p_value(0.01)
format_p_value(0.001)
format_p_value(0.0001)
format_p_value(0.0001, digits = 4)
format_p_value(0.00001)
format_p_value(0)
format_p_value(c(seq(0, 0.1, by = 0.005), 0.5, 1, NA))
format_p_value(c(seq(0, 0.1, by = 0.005), 0.5, 1, NA), digits = 2)
format_p_value(c(seq(0, 0.1, by = 0.005), 0.5, 1, NA), digits = 4)

}
