% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{get_template}
\alias{get_template}
\title{Extract template from rUrgAra}
\usage{
get_template(template_name, folder = ".")
}
\arguments{
\item{template_name}{The template to import. Possible values are : report}

\item{folder}{Folder were the files corresponding to the template should be written.}
}
\description{
Extract template from rUrgAra
}
\examples{
#DO NOT RUN
library(rUrgAra)
get_template(template_name = "report", folder = "src")
}
