% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pageGarde.R
\name{make_pageGarde}
\alias{make_pageGarde}
\title{Page de garde UrgAra}
\usage{
make_pageGarde(img_path, tab_labels, activate_showtext_auto = FALSE)
}
\arguments{
\item{img_path}{character : chemin vers la page de garde au format png ou jpg}

\item{tab_labels}{data.frame ou tibble : Tableau contenant une ligne par ligne de texte à afficher sur la page de garde et au minimum une colonnes "labels". Les colonnes "x", "y", "color", "fontfamily", "fontface" et "size" pouvent être définies mais disposent de valeurs pas défaut si omises}

\item{activate_showtext_auto}{Booléen : Utilisation de showtext::showtext_auto pour afficher des polices personnalisées. Attention si oui, l'environnement global du projet sera modifié.}
}
\value{
un plot
}
\description{
Facilite l'insertion d'une page de garde au format UrgAra dans un rapport Rmarkdown. ATTENTION : La police utilisée ne fonctionne que dans un environnement utilisant showtext::showtext_auto
}
\details{
Les paramètres de la page de garde sont anticipés pour une insertion
dans un Rmarkdown sur une page sans marge en A4
(\code{fig.width = 8.3}, \code{fig.height = 11.7}).
Les valeurs par défaut des colonnes de tab labels sont :
\itemize{
\item \code{x} : 102 (le bord droit du texte)
\item \code{y} : Espacement égal allant de y = 25 à y = 5
\item \code{color} : white
\item \code{fontface} : bold
\item \code{fontfamily} : CO_bold (cocogoose). Cette valeur nécessite d'utiliser activate_showtext_auto pour s'afficher correctement
\item \code{size} : 15
}
Ces valeurs sont projetés sur une fenetre graphique allant de 0 à 100 sur les deux axes
}
\examples{
library(rUrgAra)
tab_labels = tibble::tibble(
labels = c("Rapports TRAUMA", "Quadrimestriel", "Q1 2025",
           stringr::str_wrap("Hopital de Chambery Urgences Pédiatriques", width = 30)),
color = c("white", "white", "white", "blue"),
size  = c(25, 22, 22, 18)
)
img_path = system.file("img/exemple_page_garde.png", package = "rUrgAra")
make_pageGarde(img_path = img_path, tab_labels = tab_labels, activate_showtext_auto = TRUE)

}
