% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_add_logo.R
\name{plot_add_logo}
\alias{plot_add_logo}
<<<<<<< HEAD
\title{Add company logo to a plot}
=======
\title{add company logo to a plot}
>>>>>>> 467edd424d4727f4aa58d2bad5856ef469206fb4
\usage{
plot_add_logo(plot, logo = NULL, height = 0.1, width = 0.2)
}
\arguments{
\item{plot}{A ggplot object}

\item{logo}{The path to a .jpg logo. If NULL, the Urg'Ara logo will be used instead}

\item{height}{The height of the logo in proportion of the graph (default 0.1)}

\item{width}{The width of the logo in proportion of the graph (default 0.2)}
}
\value{
a ggplot2 object
}
\description{
Adds a company logo (by default Urg'Ara) to a plot. A custom logo can be used by specifying a path to the logo
}
\examples{
library(ggplot2)
<<<<<<< HEAD
plot_test = ggplot(data.frame(), ggplot2::aes(x = 1, y = 1)) +
=======
plot_test = ggplot(data.frame(), aes(x = 1, y = 1)) +
>>>>>>> 467edd424d4727f4aa58d2bad5856ef469206fb4
  geom_point()
plot_add_logo(plot = plot_test)
}
