% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diag_charge.R
\name{plot_diag_charge}
\alias{plot_diag_charge}
\title{Plot a charge diagram}
\usage{
plot_diag_charge(
  data,
  entry,
  exit,
  strata = NULL,
  from = NULL,
  to = NULL,
  max_LOS = Inf,
  show_lines = TRUE
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{entry}{The name of the column containing the datetime of entry. The colmum must be of class POSIXct or character in the format DD/MM/YYYY HH:MM:SS.}

\item{exit}{The name of the column containing the datetime of exit. The colmum must be of class POSIXct or character in the format DD/MM/YYYY HH:MM:SS.}

\item{strata}{The name of the colum(s) containing statas (e.g. structures) to average the charge over.}

\item{from}{The start of the time frame to consider (dmy). Number of patients will be averaged over the number of days between "from" and "to" and entry date/times will be truncated at "from". Default value is the minimum date in the entry column.}

\item{to}{The end of the time frame to consider (dmy). Number of patients will be averaged over the number of days between "from" and "to" and exit date/times will be truncated at "to". Default value is the maximum date in the exit column.}

\item{max_LOS}{Maximum length of stay for patients in minutes. Patients with durations longer than max_LOS will be considered as having missing length of stay.}

\item{show_lines}{Should the number of entry/exit be plotted ? Default is TRUE.}
}
\value{
A list. tab contains the table used to make the charge diagram. H_entry = "-1" for patients entered the day before. plot contains the charge diagram.
}
\description{
Plot a charge diagram from a table with an entre datetime and exit datetime. A charge diagram shows the number of individuals and their time of arrival for each hour/half hour of a day.
}
\examples{
library(rUrgAra)
#Table of entry/exit times
head(df_ex_charge)
#Charge diagram with exclusion of patients staying more than 3 days (72*60 = 4320 minutes)
# not taking into account strata
list_charge = plot_diag_charge(data = df_ex_charge, entry = "ENTREE",
                               exit = "SORTIE", max_LOS = 72*60)
#plot_diag_charge return two objects, a table showing for each hour of
# the day how many patient came from what hour (-1 = day before)
head(list_charge$tab)
#a charge diagram
list_charge$plot

#adding a strata to take into account that data are coming from two hospitals
list_charge_stratified = plot_diag_charge(data = df_ex_charge, entry = "ENTREE",
                                          exit = "SORTIE", strata = "Etablissement",
                                          max_LOS = 72*60)
#plot_diag_charge return two objects, a table showing for each hour of
# the day how many patient came from what hour (-1 = day before)
head(list_charge_stratified$tab)
#a charge diagram
list_charge_stratified$plot

}
