% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rUrgAra-package.R
\docType{package}
\name{rUrgAra-package}
\alias{rUrgAra}
\alias{rUrgAra-package}
\title{rUrgAra: UrgAra Package For Internal Use}
\description{
Package developped to facilitate reccurent actions for Urg'Ara users. Theses actions include applying the graphical chart to plots, creating reports and plotting a charge diagram.
}
\author{
\strong{Maintainer}: Clément Claustre \email{clement.claustre@urgences-ara.fr}

}
\keyword{internal}
