% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{UrgAra_cols}
\alias{UrgAra_cols}
\title{Extract any color from UrgAra globale palette}
\usage{
UrgAra_cols(...)
}
\arguments{
\item{...}{name of the colors to extract from a palette}
}
\value{
a vector of named HEX colors
}
\description{
Extract any color from UrgAra globale palette
}
